requirement.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
Pillow==8.0.1
requests==2.25.0
SQLAlchemy==1.3.20
urllib3==1.26.2
